treewide: resolve compiler warnings on format strings
ECICS.cpp: In function 'ECRESULT GetChanges(soap*, ECSession*, SOURCEKEY, unsigned int, unsigned int, unsigned int, unsigned int, restrictTable*, unsigned int*, icsChangesArray**)':
ECICS.cpp:1041:149: warning: format '%u' expects argument of type 'unsigned int', but argument 6 has type 'long unsigned int' [-Wformat=]
      lpSession->GetSessionManager()->GetLogger()->Log(EC_LOGLEVEL_FATAL, "%s:%d invalid size for ab entryid: %u", __FUNCTION__, __LINE__, lpDBLen[1]);
ECSessionManager.cpp: In destructor 'virtual ECSessionManager::~ECSessionManager()':
ECSessionManager.cpp:171:86: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 4 has type 'long unsigned int' [-Wformat=]
   m_lpLogger->Log(EC_LOGLEVEL_INFO, "End of session (shutdown) %llu", iSession->first);
ECSessionManager.cpp: In member function 'virtual ECRESULT ECSessionManager::CreateSession(soap*, char*, char*, char*, char*, char*, const char*, const char*, unsigned int, uint64_t, uint64_t*, ECSession**, bool, bool)':
ECSessionManager.cpp:701:91: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 5 has type 'uint64_t {aka long unsigned int}' [-Wformat=]
   m_lpLogger->Log(EC_LOGLEVEL_DEBUG, "User %s receives session %llu", szName, *lpSessionID);
ECSessionManager.cpp:703:129: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 6 has type 'uint64_t {aka long unsigned int}' [-Wformat=]
   m_lpLogger->Log(EC_LOGLEVEL_DEBUG, "User %s impersonated by %s receives session %llu", szImpersonateUser, szName, *lpSessionID);
ECSessionManager.cpp: In member function 'virtual ECRESULT ECSessionManager::CreateSessionInternal(ECSession**, unsigned int)':
ECSessionManager.cpp:766:74: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 4 has type 'uint64_t {aka long unsigned int}' [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_DEBUG, "New internal session (%llu)", newSID);
ECSessionManager.cpp: In member function 'virtual ECRESULT ECSessionManager::RemoveSession(uint64_t)':
ECSessionManager.cpp:789:78: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 4 has type 'uint64_t {aka long unsigned int}' [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_DEBUG, "End of session (logoff) %llu", sessionID);
ECSessionManager.cpp: In static member function 'static void* ECSessionManager::SessionCleaner(void*)':
ECSessionManager.cpp:938:104: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 4 has type 'long unsigned int' [-Wformat=]
     lpSessionManager->m_lpLogger->Log(EC_LOGLEVEL_INFO, "End of session (timeout) %llu", iRemove->first);
ECSessionManager.cpp: In member function 'ECRESULT ECSessionManager::DumpStats()':
ECSessionManager.cpp:1442:129: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 5 has type 'ULONGLONG {aka long unsigned int}' [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_FATAL, "  Sessions : %u (%llu bytes)", sSessionStats.session.ulItems, sSessionStats.session.ullSize);
ECSessionManager.cpp:1444:125: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 5 has type 'ULONGLONG {aka long unsigned int}' [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_FATAL, "  Groups   : %u (%llu bytes)" ,sSessionStats.group.ulItems, sSessionStats.group.ullSize);
ECSessionManager.cpp:1452:140: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 5 has type 'ULONGLONG {aka long unsigned int}' [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_FATAL, "  Open tables: %u (%llu bytes)", sSessionStats.session.ulOpenTables, sSessionStats.session.ulTableSize);
ECSessionManager.cpp:1460:85: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 4 has type 'ULONGLONG {aka long unsigned int}' [-Wformat=]
  m_lpLogger->Log(EC_LOGLEVEL_FATAL, "  Mem usage : %llu Bytes", sSearchStats.ullSize);
ECNotificationManager.cpp: In member function 'HRESULT ECNotificationManager::AddRequest(uint64_t, soap*)':
ECNotificationManager.cpp:146:98: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 4 has type 'uint64_t {aka long unsigned int}' [-Wformat=]
   m_lpLogger->Log(EC_LOGLEVEL_WARNING, "Replacing notification request for ID: %llu", ecSessionId);
ECExchangeExportChanges.cpp: In member function 'HRESULT ECExchangeExportChanges::ExportMessageChangesFast()':
../../common/include/zarafa/ECLogger.h:91:47: warning: format '%u' expects argument of type 'unsigned int', but argument 5 has type 'std::vector<ICSCHANGE>::size_type {aka long unsigned int}' [-Wformat=]
    (_plog)->Log(EC_LOGLEVEL_DEBUG, __VA_ARGS__); \
ECExchangeExportChanges.cpp:1297:2: note: in expansion of macro 'ZLOG_DEBUG'
  ZLOG_DEBUG(m_lpLogger, "ExportFast: At step %u, changeset contains %u items)", m_ulStep, m_lstChange.size());

+++ provider/client/ECExchangeExportChanges.cpp
@@ -1294,7 +1294,8 @@ HRESULT ECExchangeExportChanges::ExportMessageChangesFast()
 	LPSPropTagArray lpImportProps = m_sourcekey.empty() ? (LPSPropTagArray)&sptImportPropsServerWide : (LPSPropTagArray)&sptImportProps;
 
 	// No more changes (add/modify).
-	ZLOG_DEBUG(m_lpLogger, "ExportFast: At step %u, changeset contains %u items)", m_ulStep, m_lstChange.size());
+	ZLOG_DEBUG(m_lpLogger, "ExportFast: At step %u, changeset contains %lu items)",
+		m_ulStep, static_cast<unsigned long>(m_lstChange.size()));
 	if (m_ulStep >= m_lstChange.size())
 		goto exit;
